Added "attached" bool property to tempo-nodes.
The purpose of this property would be to attach this tempo node
(its beginning) to all tracks. Thus tempo changes before the
attached node move all tracks along with the attached node.
The property already gets toggled upon double click on the tempo
node and causes the node to become filled/empted rectangle.
Loading and saving of the property is implemented.