Support changing a number of bars of a tempo-node.
A bars-scrollbox was added to the TimeScaleForm. Its value is
set to number of bars of the currently selected item (tempo-node).
If a number of bars changes, the enabled actions (i.e. Add, Update)
take this new value into account.
Updating bars of a node preserves tempo for nodes before and after
the attached node. However, changing bars of the attached node
followed by another attached node recalculates the node-tempo so
that bars fit into existing frames of both tempo nodes.